refactor(ParseFutureBuilder): Self manage state
Instead of managing state just inside the retryFutureBuilder,
the state of the queryFunction is saved inside ParseFutureBuilder.

This is a workaround for not triggering the future function every time
the widget is rebuilt. This was the case, because the queryFunction
would be wrapped in a class function. This function was being rebuilt
every time the ParseFutureBuilder was rebuild, therefore
the RetryFutureBuilder would retrigger the future.

Now the future is only reloaded, if the queryFunction object changes.
Reloading can now be triggered by the caller of ParseFutureBuilder
by setting a new queryFunction